fixed Login & Remember Me Function

+++ script.js
@@ -31,10 +31,12 @@ async function loadData() {
   await loadUsers();
   if (localStorage.getItem("userData") != null) {
     let data = JSON.parse(localStorage.getItem("userData"));
-    if(!data.remberMe) {
+    if(data.loginCount == 2 && !data.remberMe) {
       localStorage.clear();
       window.location.href = "/assets/templates/login/login.html";
     }
+    data.loginCount++;
+    localStorage.setItem("userData", JSON.stringify(data));
     await loadDataToUser();
   } else if (
     !window.location.href.match("/assets/templates/login/login.html") &&
